Split ability test class into separate class for each role
